Re-org, cleanup, documentation
